Moved cachedTokenStorerAndGetter to new file
